fix debian_package_with_docker on Mac
With docker-machine it's not possible to use docker volumes
outside of /Users. This means that the two volumes used for the
.deb creation (a temporary directory in /private/tmp/... and the
final package directory in /tmp) need to use different
directories, currently in the same directory as the Doit source
code itself.

The possibly created package directory is now listed in
.gitignore.

While doing the change, Cwd.pm is now always loaded via use and
not on-demand via require.