Doit::Sudo: reset terminal to sane settings after start of sudo worker
Recent versions of sudo per default set the use_pty option in
/etc/sudoers. This has negative effects on the current terminal
and it also fails to reset the terminal settings after the end
of Doit scripts. To workaround this, Doit::Sudo::do_connect waits
until the sudo worker is settled (by sending a ping to the worker)
and tries to change the terminal to sane settings. This is
preferably done using "stty sane". If this is not possible or fails,
then a fallback to POSIX::Termios is done.